+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Update using ```composer update```.
 * Added multi database support to migrations.
 * Added an option to opt out of atomic `getOrElse` for APCU caching. Use the `atomic_get_or_else` key.
 * Added `ConnectionManager::close()` method.
+* Added `ConnectionManager::executeAndClose()` method.
 
 #### Changes